+++ spec/models/usage_report_spec.rb
@@ -11,6 +11,7 @@ RSpec.describe UsageReport do
       s4 = create(:service, status: :draft)
       create(:service, status: :published)
 
+      create(:offer, service: s1, order_type: :order_required)
       create(:offer, service: s1, order_type: :order_required, status: :draft)
       create(:offer, service: s1, order_type: :open_access, status: :published)
       create(:offer, service: s1, order_type: :order_required, order_url: "http://order.com", status: :published)
@@ -25,7 +26,7 @@ RSpec.describe UsageReport do
   context ".not_orderable_count" do
     it "counts open access and external services" do
       s1 = create(:service, status: :published)
-      s2 = create(:service, status: :published)
+      s2 = create(:open_access_service, status: :published)
       s3 = create(:service, status: :unverified)
       s4 = create(:service, status: :draft)
       create(:service, status: :published)