+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 
 ### Fixed
 - Redirection after Scientific Domain deletion in backoffice (@kmarszalek)
+- Consistence of Resource/Offer order types (@goreck888)
 
 ### Security